requirements.txt
Django==4.2.13
python-slugify==8.0.4
